Create a MainCLI fn with utilities
