+++ examples/cli/example.toml
@@ -1,0 +1,12 @@
+[input]
+name = "List"
+
+	[input.config]
+	files=["testdata/input.csv.zst"]
+
+[output]
+name = "FileWriter"
+procs=1
+
+	[output.config]
+	PathString="./_out/output.csv.gz"
